requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pip==23.0.1
pyTelegramBotAPI==4.10.0
requests==2.28.2
setuptools==65.5.0
urllib3==1.26.15
pydantic==1.10.6
python-dotenv==1.0.0
peewee==3.16.0
loguru==0.6.0
translate==3.6.1